+++ app/src/main/java/com/chico/myhomebookkeeping/obj/ConstantsOfUpdate.kt
@@ -7,4 +7,6 @@ object ConstantsOfUpdate {
     const val UPDATE_44_TO_45 = "update_program_44_to_45"
     const val UPDATE_59_TO_60 = "update_program_59_to_60"
     const val UPDATE_ADD_SICK_LEAVE_INCOME_CATEGORY = "update_add_sick_leave_income_category"
+    const val UPDATE_69_TO_70_ADD_SUBSCRIPTIONS_ONLINE_SERVICES_CATEGORY =
+        "update_program_69_to_70_add_subscriptions_online_services_category"
 }